Hooked up VIN decode to UI, fixed some bugs with add and remove not working properly. Adding a vehicle now updates both server and settings list.
Version 0.5.5 Build 27

+++ app/src/main/java/com/aspace/aspace/retrofitmodels/Car.java
@@ -26,7 +26,7 @@ public class Car {
     private String name;
     @SerializedName("length")
     @Expose
-    private Integer length;
+    private Double length;
 
     public String getId() {
         return id;
@@ -76,11 +76,11 @@ public class Car {
         this.name = name;
     }
 
-    public Integer getLength() {
+    public Double getLength() {
         return length;
     }
 
-    public void setLength(Integer length) {
+    public void setLength(Double length) {
         this.length = length;
     }
 }
